chore: domain에서 article/comment 패키지 분리 (#47)
