--- src/main/java/com/soda/article/domain/CommentUpdateRequest.java
@@ -1,16 +1,0 @@
-package com.soda.article.domain;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class CommentUpdateRequest {
-
-    private String content;
-
-}
