+++ src/main/java/com/soda/article/domain/article/ArticleViewResponse.java
@@ -1,6 +1,7 @@
-package com.soda.article.domain;
+package com.soda.article.domain.article;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
+import com.soda.article.domain.comment.CommentDTO;
 import com.soda.article.enums.PriorityType;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
