+++ src/main/java/com/soda/article/domain/comment/CommentUpdateRequest.java
@@ -1,4 +1,4 @@
-package com.soda.article.domain;
+package com.soda.article.domain.comment;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -6,10 +6,11 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Getter
-@Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class ArticleFileDTO {
-    private String name;
-    private String url;
+@Builder
+public class CommentUpdateRequest {
+
+    private String content;
+
 }
