+++ src/main/java/com/soda/request/dto/response/ResponseDTO.java
@@ -1,9 +1,6 @@
 package com.soda.request.dto.response;
 
-import com.soda.article.domain.ArticleDTO;
 import com.soda.request.dto.link.LinkDTO;
-import com.soda.request.dto.request.RequestDTO;
-import com.soda.request.entity.Request;
 import com.soda.request.entity.Response;
 import com.soda.request.enums.RequestStatus;
 import lombok.Builder;
